Fix bug in quantile: where the input probability is given as a float.
In Pharo 9, in cases like where the input probability is a Float and not a Fraction like
```
#(1 2 3 4 5) asSortedCollection quantile: 0.25
```
an error gets thrown as SortedCollection requires an Integer for a lookup. With the previous code
the value of the temporary value, `p`, could end up being a Float like `2.0` that would equal the integer
( i.e. `p truncate = p` ) but using `2.0` as an index no longer works.